requirements.txt
Flask>=2.0
flask-cors>=3.0
google-generativeai>=0.2.0
